No need for new module at all.
