+++ maven-resolver-spi/src/main/java/org/eclipse/aether/spi/connector/transport/http/ChecksumExtractor.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.eclipse.aether.transport.shared.http;
+package org.eclipse.aether.spi.connector.transport.http;
 
 import java.util.Map;
 import java.util.function.Function;
